Implemented viewport as a concept within grid views, and coordinate space translators for viewports.  Re-organized namespaces to better reflect current state.
